requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
langchain==1.2.13
langchain-openai==1.1.11
langgraph==1.1.3
langgraph-checkpoint-redis==0.4.0
langgraph-checkpoint-postgres==3.0.5
chromadb==1.5.5
python-dotenv==1.2.2
httpx==0.28.1
redis==7.3.0
PyJWT[crypto]==2.10.1
cryptography==46.0.5
psycopg[binary]==3.3.3
pytest==8.4.2
mcp==1.26.0
psycopg2-binary==2.9.11
asyncpg==0.31.0
sqlalchemy==2.0.48
streamlit==1.55.0
requests==2.32.5


